+++ FirebaseAuth/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 11.1.0
+- [fixed] Fixed `Swift.error` conformance for `AuthErrorCode`. (#13430)
+- [added] Added custom provider support to `AuthProviderID`. Note that this change will be breaking
+  to any code that implemented an exhaustive `switch` on `AuthProviderID` in 11.0.0 - the `switch`
+  will need expansion. (#13429)
+
 # 11.0.0
 - [fixed] Fixed auth domain matching code to prioritize matching `firebaseapp.com` over `web.app`
   even if the server returns the `web.app` domain listed first. (#7992)